requirements.txt
google-generativeai
yt-dlp
line-bot-sdk
python-dotenv